requirements.txt
astroid==3.3.8
beautifulsoup4==4.13.1   
colorama==0.4.6
dill==0.3.9
flake8==7.1.1
isort==6.0.0
lxml==5.3.0
Markdown==3.4.3
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.0.0   
platformdirs==4.3.6      
pycodestyle==2.12.1      
pyflakes==3.2.0
pylint==3.3.4
python-docx==0.8.11      
soupsieve==2.6
tomlkit==0.13.2
typing_extensions==4.12.2